Split IntelliSense providers
