+++ src/common/utility.ts
@@ -169,8 +169,7 @@ export class Utility {
     }
 
     public static expandModules(input: string, moduleMap: Map<string, string>): string {
-        const pattern: RegExp = new RegExp(/\${MODULES\..+}/g);
-        return input.replace(pattern, (matched) => {
+        return input.replace(Constants.imagePlaceholderPattern, (matched) => {
             const key: string = matched.replace(/\$|{|}/g, "");
             if (moduleMap.has(key)) {
                 const value: string = moduleMap.get(key);
@@ -264,5 +263,14 @@ export class Utility {
                 imageToDockerfileMap.set(image, path.join(modulePath, module.image.tag.platforms[platform]));
             });
         }
+    }
+
+    // Remove the wrapping "${" and "}" of a image placeholder
+    public static unwrapImagePlaceholder(imagePlaceholder: string): string {
+        if (imagePlaceholder.search(Constants.imagePlaceholderPattern) === 0 && imagePlaceholder.endsWith("}")) {
+            return imagePlaceholder.slice(2, -1);
+        }
+
+        return undefined;
     }
 }
