+++ src/intelliSense/configDefinitionProvider.ts
@@ -1,0 +1,18 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license.
+
+"use strict";
+import * as vscode from "vscode";
+import { IntelliSenseUtility } from "./intelliSenseUtility";
+
+export class ConfigDefinitionProvider implements vscode.DefinitionProvider {
+    public async provideDefinition(document: vscode.TextDocument, position: vscode.Position): Promise<vscode.Definition> {
+        const {dockerfile, range} = await IntelliSenseUtility.getImageDockerfileAtLocation(document, position);
+        if (dockerfile) {
+            const beginOfFile: vscode.Position = new vscode.Position(0, 0);
+            return new vscode.Location(vscode.Uri.file(dockerfile), beginOfFile);
+        }
+
+        return undefined;
+    }
+}
